test_fuse_elewise_add_act_pass
